moved verify to inner
